feat: print cost and memory analysis of train_step
